Crate arrabiata

Modules

column_env

This module will be used by the prover to evaluate at a certain point.

columns

constraints

interpreter

This module contains the implementation of the IVC scheme in addition to running an arbitrary function that can use up to crate::NUMBER_OF_COLUMNS columns. At the moment, all constraints must be of maximum degree crate::MAX_DEGREE, but it might change in the future.

logup

This file will implement a logup argument to allow users performing lookup in their circuits.

poseidon_3_60_0_5_5_fp

poseidon_3_60_0_5_5_fq

proof

prover

A prover for the folding/accumulation scheme

verifier

A verifier for the folding/accumulation scheme

witness

Constants

IVC_CIRCUIT_SIZE

The number of rows the IVC circuit requires.

MAXIMUM_FIELD_SIZE_IN_BITS

The maximum number of bits the fields can be. It is critical as we have some assumptions for the gadgets describing the IVC.

MAX_DEGREE

The maximum degree of the polynomial that can be represented by the polynomial-time function the library supports.

MIN_SRS_LOG2_SIZE

The minimum SRS size required to use Nova, in base 2. Requiring at least 2^16 to perform 16bits range checks.

NUMBER_OF_COLUMNS

The maximum number of columns that can be used in the circuit.

NUMBER_OF_PUBLIC_INPUTS

The maximum number of public inputs the circuit can use per row We do have 15 for now as we want to compute 5 rounds of poseidon per row using the gadget crate::columns::Gadget::Poseidon. In addition to the 12 public inputs required for the rounds, we add 2 more for the values to absorb.

NUMBER_OF_SELECTORS

The number of selectors used in the circuit.

NUMBER_OF_VALUES_TO_ABSORB_PUBLIC_IO

Define the number of values we must absorb when computating the hash to the public IO.

POSEIDON_ALPHA

The low-exponentiation value used by the Poseidon hash function for the substitution box.

POSEIDON_ROUNDS_FULL

The number of full rounds in the Poseidon hash function.

POSEIDON_STATE_SIZE

The number of elements in the state of the Poseidon hash function.